Move the pet search files to a separate directory
